requirements.txt
flask
flask-login
mysql-connector-python
psycopg2-binary
gunicorn
requests